Update portscan.cpp
Implemented some basic skeleton code in the main function; starts with a code block to be incorporated for taking both command line inputs or user input.
Still missing code for if-statement in line 28.
Is checking for valid regex match in line 35 done? I looked it up and what's already there seems correct.